Simple implementation. All test passed.

+++ src/main/java/ru/mail/polis/fedoseenkov/MyService.java
@@ -7,6 +7,7 @@ import ru.mail.polis.KVService;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
+import java.util.NoSuchElementException;
 
 public class MyService implements KVService {
     private static final String PREFIX = "id=";
@@ -18,7 +19,7 @@ public class MyService implements KVService {
 
     @NotNull
     private static String extractId(@NotNull final String query){
-          if (query.startsWith(PREFIX)) {
+          if (!query.startsWith(PREFIX)) {
             throw new IllegalArgumentException("bad string");
         }
 
@@ -47,33 +48,46 @@ public class MyService implements KVService {
                 "/v0/entity",
                 http -> {
                     final String id = extractId(http.getRequestURI().getQuery());
+                    if (id.equals("")) http.sendResponseHeaders(400, 0);
+                    else {
+                        switch (http.getRequestMethod()) {
+                            case "GET":
+                                final byte[] getValue;
+                                try {
+                                    getValue = dao.get(id);
+                                } catch (NoSuchElementException | IOException e) {
+                                    http.sendResponseHeaders(404, 0);
+                                    break;
+                                }
+                                http.sendResponseHeaders(200, getValue.length);
+                                http.getResponseBody().write(getValue);
+                                break;
+                            case "DELETE":
+                                dao.delete(id);
+                                http.sendResponseHeaders(202, 0);
+                                break;
+                            case "PUT":
+                                final int contentLenght = http.getRequestBody().available();
+                                if (contentLenght == 0) {
+                                    dao.upsert(id, new byte[0]);
+                                    http.sendResponseHeaders(201, 0);
+                                    break;
+                                }
+                                else {
+                                    final byte[] putValue = new byte[contentLenght];
+                                    if (http.getRequestBody().read(putValue) != putValue.length) {
+                                        throw new IOException("Can't read file at once.");
+                                    }
 
-                    switch (http.getRequestMethod()) {
-                        case "GET":
-                            final byte[] getValue = dao.get(id);
-                            http.sendResponseHeaders(200, getValue.length);
-                            http.getResponseBody().write(getValue);
-                            break;
-                        case "DELETE":
-                            dao.delete(id);
-                            http.sendResponseHeaders(202, 0);
-                            break;
-                        case "PUT":
-                            final int contentLenght =
-                                    Integer.valueOf(
-                                            http.getRequestHeaders().getFirst("Content-Lenght"));
-                            final byte[] putValue = new byte[contentLenght];
-                            if (http.getRequestBody().read(putValue) != putValue.length) {
-                                throw new IOException("Can't read file at once.");
-                            }
-
-                            dao.upsert(id, putValue);
-                            http.sendResponseHeaders(201, 0);
-                            break;
-                        default:
-                            http.sendResponseHeaders(404, 0);
+                                    dao.upsert(id, putValue);
+                                    http.sendResponseHeaders(201, 0);
+                                    break;
+                                }
+                            default:
+                                http.sendResponseHeaders(405, 0);
+                                break;
+                        }
                     }
-
                     http.close();
                 });
     }
